pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "migsafe"
version = "0.4.0"
description = "Safe migrations for production PostgreSQL databases"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Kirill Vladimirov"}
]
keywords = [
    "alembic",
    "migrations",
    "postgresql",
    "database",
    "sql",
    "django",
    "ci",
    "lint",
    "safety",
    "production",
    "ast",
    "analysis",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "GitPython>=3.1.0",
    "typing_extensions>=4.0.0; python_version<'3.10'",
]

[project.urls]
Homepage = "https://github.com/SuccubHunter/MigrationSafe"
Documentation = "https://github.com/SuccubHunter/MigrationSafe#readme"
Repository = "https://github.com/SuccubHunter/MigrationSafe"
"Bug Tracker" = "https://github.com/SuccubHunter/MigrationSafe/issues"
Changelog = "https://github.com/SuccubHunter/MigrationSafe/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
]
formatters = [
    "rich>=13.0",  # для улучшенного текстового вывода
]
executors = [
    "psycopg2-binary>=2.9.0",  # для работы с PostgreSQL
    "alembic>=1.8.0",  # для выполнения миграций
    "sqlalchemy>=1.4.0",  # для парсинга URL подключения
]

[project.scripts]
migsafe = "migsafe.cli:main"

[tool.setuptools]
packages = ["migsafe", "migsafe.analyzers", "migsafe.formatters", "migsafe.rules", "migsafe.sources", "migsafe.autofix", "migsafe.executors", "migsafe.stats", "migsafe.history"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt"]

# Конфигурация для black
[tool.black]
line-length = 127
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# Конфигурация для isort
[tool.isort]
profile = "black"
line_length = 127
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*", "*/__pycache__/*"]

# Конфигурация для mypy
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
ignore_missing_imports = true
follow_imports = "normal"
exclude = [
    "^build/",
    "^dist/",
    "^migsafe\\.egg-info/",
]

[[tool.mypy.overrides]]
module = [
    "git.*",
    "click.*",
    "pydantic.*",
]
ignore_missing_imports = true

# Конфигурация для Ruff
[tool.ruff]
line-length = 127
target-version = "py38"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/integration/fixtures/django_migrations/*.py" = ["N999"]

[tool.ruff.lint.isort]
known-first-party = ["migsafe"]

# Конфигурация для flake8 (через setup.cfg или .flake8)
# flake8 не поддерживает pyproject.toml напрямую, создадим .flake8
